Improve province selection
Fix a slight timing bug with province selection, where if a click was executed with a slight drag into a different province, the second province would be selected instead of the first.
Fix a small bug with the province name label and the province history button when using multi-selection to completely deselect all provinces.